refactor : 최종 상금 계산 로직 계산기 따로 생성
