+++ src/main/java/lotto/domain/lottos/user/WinningLotto.java
@@ -2,13 +2,15 @@ package lotto.domain.lottos.user;
 
 import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 import lotto.domain.Rank;
+import lotto.domain.calculators.FinalPrizeCalculator;
 
 public class WinningLotto {
+    private final FinalPrizeCalculator finalPrizeCalculator;
     private final EnumMap<Rank, Integer> ranks = new EnumMap<>(Rank.class);
 
-    public WinningLotto() {
+    public WinningLotto(FinalPrizeCalculator finalPrizeCalculator) {
+        this.finalPrizeCalculator = finalPrizeCalculator;
         initRanks();
     }
 
@@ -18,32 +20,18 @@ public class WinningLotto {
         }
     }
 
+    private long calculateTotalPrizeAmount() {
+        return finalPrizeCalculator.calculate(ranks);
+    }
+
     public long getFinalPrizeAmount() {
         return calculateTotalPrizeAmount();
-    }
-
-    private long calculateTotalPrizeAmount() {
-        long result = 0;
-
-        for (Map.Entry<Rank, Integer> entry : ranks.entrySet()) {
-            Rank rank = entry.getKey();
-            int matchedCount = entry.getValue();
-            long prizeMoney = rank.getPrizeMoney();
-
-            if (isMatchedRank(matchedCount)) {
-                result += (prizeMoney * matchedCount);
-            }
-        }
-        return result;
-    }
-
-    private boolean isMatchedRank(int matchedCount) {
-        return matchedCount > 0;
     }
 
     public EnumMap<Rank, Integer> getWinningStatistics() {
         return ranks;
     }
+
 
     private void initRanks() {
         for (Rank rank : Rank.values()) {
